% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Efficiently convert selected columns of \code{data.table} to 
nquad statements}
\usage{
convert(dat, schema_list, type = "", datatype = "", done_so_far = 0)
}
\arguments{
\item{type}{Type of nquads statements: \code{'uri'} or \code{'literal'}.}

\item{df}{Name of the data.table as a length one character vector.}

\item{column_names}{Names of the data.table to be converted as a character vector.}

\item{base}{Base of the URIs to be minted in \code{uriref},}

\item{paths}{Further path of URI to be minted in \code{uriref}.}

\item{datatypes}{Character vector containing the datatype of literals. May include language statements.}
}
\description{
Efficiently convert selected columns of \code{data.table} to 
nquad statements
}
\examples{
the_data = data.table(refs = c('a', 'b', 'c'), ltrl = c(1, 2, 3))
convert(df = "the_data", column_names = "ltrl", type = 'literal', datatypes = 'xsd:int')
convert(df = "the_data", column_names = "refs", type = 'uri', base = 'https://www.example.org/', paths = 'data/')
print(the_data)
}
