% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cower-package.R
\docType{package}
\name{cower-package}
\alias{cower-package}
\title{The cower Package}
\description{
Power Analysis for Correlation Coefficients.
}
\details{
\tabular{ll}{
Package: \tab cower\cr
Type: \tab Package\cr
Version: \tab 0.0.02\cr
Date: \tab 2018-07-06\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL-2 + file LICENSE\cr
LazyLoad: \tab yes\cr
URL: \tab https://github.com/m-Py/cower\cr
}

Power analysis for the comparison of independent correlation coefficients.
}
\author{
Martin Papenberg <martin.papenberg@hhu.de>
}
\keyword{package}
