% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cower.R
\name{cower}
\alias{cower}
\title{Convert a csv file to nquads using json-ld data description.}
\usage{
cower(csv_path, json_path, nquad_path, compress = TRUE, batch_size = -1L,
  max_size = FALSE)
}
\arguments{
\item{csv_path}{The path to the csv file.}

\item{json_path}{The path to the json-ld data description. 
Use \code{\link{build_schema_list}} and \code{\link{schema_json}} to make one or use COW.}

\item{nquad_path}{Destination path for nquads. \code{cower} will
give a warning if the extension does not match the \code{compress} option.}

\item{batch_size}{Batch size for conversion. Default is \code{-1L},
all rows at once. For large csv-files that take up considerable memory
to convert, setting this to a lower number will preserve memory.}

\item{max_size}{Set this parameter to an integer to only convert 
part of the dataset. Useful for testing. Depends on \code{head} command line tool.}

\item{compress.}{Should cower output be a compressed (gz) file?
Defaults to \code{TRUE} as nquad files can be very large relative
to the original file (x10, roughly).}
}
\description{
Convert a csv file to nquads using json-ld data description.
}
