% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_names.R
\name{graph_names}
\alias{graph_names}
\title{Creates URIs for graphs}
\usage{
graph_names(filehash, base = "")
}
\arguments{
\item{base}{The base URI to be used in the rdf.}

\item{path_to_csv}{The path to the csv file.}

\item{sub_path}{Further path of the URI.}
}
\value{
A named vector with five graph names for the 
     assertion, nanopublication, provenance, pubinfo, and
     resource graphs.
}
\description{
\code{graph_names} creates URIs for graphs based on a file
from file hash and system time.
}
\examples{
path = tempfile()
outfile = file(path, "wb")
write.csv(data.frame(a = c(1, 2, 3)), file = outfile)
close(outfile)
graph_names(path)
}
