% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{literal}
\alias{literal}
\title{Create literals}
\usage{
literal(string, datatype = "xsd:string", ...)
}
\arguments{
\item{string}{The value to be converted into a literal.}

\item{datatype}{The datatype of the literal. Defaults to xsd:string. Language tags are identified by \code{@}.}
}
\value{
A literal with escaped quotes and the datatype appended. If \code{string} is \code{NA}, \code{NA} is returned.
}
\description{
\code{literal} creates a literal from a value with escaped quotes for file writing and appends a datatype.
}
\examples{
literal("example", datatype="xsd:string")
literal("example", datatype="@en")
literal(630.6, datatype="xsd:float")
literal(NA, datatype="xsd:integer")
}
