% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.indep.cor.R
\name{power.indep.cor}
\alias{power.indep.cor}
\title{Power analysis for two independent correlations}
\usage{
power.indep.cor(r1, r2, n1 = NULL, n2 = NULL, power = NULL,
  sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{r1}{Value of first Pearson correlation coefficient, must be passed.}

\item{r2}{Value of second correlation coefficient, must be passed.}

\item{n1}{Sample size of group 1; pass only for 'post hoc' analysis.}

\item{n2}{Sample size of group 2; pass only for 'post hoc' analysis.}

\item{power}{The statistical power; pass only for 'a-priori' analysis.}

\item{sig.level}{The employed alpha level, defaults to 0.05.}

\item{alternative}{Must be 'greater', 'less', or 'two.sided'. 'greater' (r1 > r2) or 'less' (r1 < r2) will result in one sided tests (to be used if a directional hypothesis exists). Default test is two.sided.}
}
\value{
A \code{list} that contains all passed and computed values.
  \item{r1}{passed correlation coefficient r1}
  \item{r2}{passed correlation coefficient r2}
  \item{q}{The effect size. Represents the difference between the z-standardized values of r1 and r2}
  \item{n1}{sample size in group 1, either passed ('post hoc'), or computed ('a-priori')}
  \item{n2}{sample size in group 1, either passed ('post hoc'), or computed ('a-priori')}
  \item{power}{Power to detect a significant difference between the two correlation coefficients r1 and r2, either passed ('a-priori') or computed ('post hoc')}
  \item{sig.level}{The significance level that was employed for the power analysis}
  \item{hypothesis}{Passed argument 'alternative'}
}
\description{
This function can be used to conduct power analyses for the comparison of independent correlation coefficients, as tested via Fisher's z-test. It can (1) determine the power for detecting a significant difference between 2 correlations given N ('post hoc'), and (2) determine the required N to achieve a desired power ('a-priori').
}
\examples{
power.indep.cor(r1 = 0.4, r2 = 0.3, power = 0.8)

power.indep.cor(r1 = 0.4, r2 = 0.3, n1 = 450, n2 = 450, alternative = "greater")
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
\references{
Fisher RA. \emph{Statistical Methods for Research Workers}. Edinburgh, Scotland: Oliver and Boyd; 1925. Available: http://psychclassics.yorku.ca.

   Cohen, J. (1969). \emph{Statistical power analysis for the behavioural sciences}. New York: Academic Press.
}

