% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr_cor-package.R
\docType{package}
\name{pwr_cor-package}
\alias{pwr_cor-package}
\title{The pwr_cor Package}
\description{
Power Analysis for Correlation Coefficients.
}
\details{
\tabular{ll}{
Package: \tab pwr_cor\cr
Type: \tab Package\cr
Version: \tab 0.01-1\cr
Date: \tab 2015-08-20\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Conduct power analyses for correlation coefficients
}
\author{
Martin Papenberg
}
\keyword{package}

