% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_rdf}
\alias{read_rdf}
\title{Read rdf file and create a data.table.}
\usage{
read_rdf(path_to_rdf, type = c("ntriples", "nquads"))
}
\arguments{
\item{path_to_rdf}{The path of the file.}

\item{type}{the rdf serialisation, either 'ntriples' or 'nquads'}
}
\description{
Read rdf file and create a data.table.
}
