% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{uriref}
\alias{uriref}
\title{Create URI}
\usage{
uriref(string, base, path = "", ...)
}
\arguments{
\item{string}{The value to be converted to a URI.}

\item{base}{The base of URI.}

\item{path}{Further path of URI}
}
\description{
\code{uriref} creates a uri with nquad-angle brackets and any non-
safe characters are %-encoded or replaced by underscores.
}
\examples{
uriref('rainy', base = 'http://www.weather.com/', path = 'weather/')
}
