% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_schema.R
\name{build_schema_list}
\alias{build_schema_list}
\title{Create a basic json-ld metadata from a csv dataset.}
\usage{
build_schema_list(csvpath, delimiter = ",", encoding = "UTF-8",
  base = "https://iisg.amsterdam/resource/", dataset_name = "")
}
\description{
Create a basic json-ld metadata from a csv dataset.
}
