% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.indep.cor.R
\name{power.indep.cor}
\alias{power.indep.cor}
\title{Power analysis for two independent correlations}
\usage{
power.indep.cor(r1, r2, n1 = NULL, n2 = NULL, power = NULL,
  sig.level = 0.05, alternative = "two.sided")
}
\arguments{
\item{r1}{Value of first Pearson correlation coefficient}

\item{r2}{Value of the other Pearson correlation coefficient}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}

\item{sig.level}{The employed alpha level}

\item{alternative}{Must be either 'greater', 'less', or 'two.sided'. 'greater' or 'less' will result in one sided tests (to be used if a directional hypothesis exists), two.sided in a two sided test.}
}
\value{
Single value \code{Vector}. The power to detect a difference between the two given correlation coefficients.
}
\description{
This function determines the power for detecting a significant difference between two independent Pearson correlation coefficients given n for each group. The power is determined for a Fischer z-test. Currently, it does not behave like functions from package `pwr`. It can only be used to determine the power given two r; not n, effect (q) or significance level.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
\references{
Fisher RA. \emph{Statistical Methods for Research Workers}. Edinburgh, Scotland: Oliver and Boyd; 1925. Available: http://psychclassics.yorku.ca.

  Cohen, J. (1969). \emph{Statistical power analysis for the behavioural sciences}. New York: Academic Press.
}

